pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ci-cd-security-policy-gate"
version = "0.1.0"
description = "A Python CLI tool that scans GitHub repositories for insecure CI/CD configurations and enforces policy gates."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Oli Bonnick" }
]
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]

[project.scripts]
policy-gate = "policy_gate.cli:main"

[tool.setuptools.packages.find]
include = ["policy_gate*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
